requirements.txt
bitarray==2.5.1
tabulate==0.8.9
pika==1.3.0
pycryptodome==3.15.0
Command==0.1.0